api: split SerialisedScript to internal/common
