+++ plutus-ledger-api/plutus-ledger-api.cabal
@@ -85,6 +85,7 @@ library
   other-modules:
     PlutusLedgerApi.Internal.ParamName
     PlutusLedgerApi.Internal.EvaluationContext
+    PlutusLedgerApi.Internal.SerialisedScript
 
     Codec.CBOR.Extras
     Prettyprinter.Extras
