+++ plutus-ledger-api/src/PlutusLedgerApi/Common.hs
@@ -1,6 +1,6 @@
 -- | Common types and functions used across all the ledger API modules.
 module PlutusLedgerApi.Common
-    -- NOTE: there is a reason for not using the `as Export` trick: it makes haddock ugly/verbose
+    -- There is a reason for not using the `as Export` trick: it makes haddock ugly/verbose
     ( module PlutusLedgerApi.Common.Eval
     , module PlutusLedgerApi.Common.SerialisedScript
     , module PlutusLedgerApi.Common.PlutusVersions
